feat: 홈화면 - 매장 상세정보 조회 (DURI-220)

+++ app/src/main/java/kr/com/duri/groomer/domain/entity/Shop.java
@@ -62,7 +62,7 @@ public class Shop extends BaseEntity {
     private String kakaoTalk; // 오픈카톡방 링크
 
     @Column(name = "shop_rating")
-    private Integer rating; // 매장 별점
+    private Float rating; // 매장 별점
 
     @Builder.Default
     @Column(name = "new_shop")
